[src] Move all code related to articulated device inside from BaseDevice to articulatedDeviceController
