+++ src/SofaHapticAvatar/HapticAvatar_BaseDeviceController.h
@@ -7,17 +7,9 @@
 #pragma once
 
 #include <SofaHapticAvatar/config.h>
-
-#include <sofa/defaulttype/RigidTypes.h>
 #include <sofa/type/Vec.h>
 
 #include <SofaUserInteraction/Controller.h>
-#include <SofaHaptics/LCPForceFeedback.h>
-
-#include <SofaHapticAvatar/HapticAvatar_DriverPort.h>
-#include <SofaHapticAvatar/HapticAvatar_PortalManager.h>
-#include <SofaHapticAvatar/HapticAvatar_HapticThreadManager.h>
-
 
 namespace sofa::HapticAvatar
 {
@@ -33,6 +25,10 @@ class SOFA_HAPTICAVATAR_API HapticAvatar_BaseDeviceController : public Controlle
 public:
     SOFA_CLASS(HapticAvatar_BaseDeviceController, Controller);
 
+    using Coord = Vec1Types::Coord;
+    using VecCoord = Vec1Types::VecCoord;
+    using VecDeriv = Vec1Types::VecDeriv;
+
     /// default constructor
     HapticAvatar_BaseDeviceController();
 
@@ -42,38 +38,29 @@ public:
     /// Component API 
     ///{
     void init() override;
-    void bwdInit() override;    
     void handleEvent(core::objectmodel::Event *) override;
     void draw(const sofa::core::visual::VisualParams* vparams) override;
     ///}
-
-
-    HapticAvatar_DriverPort* getHapticDriver() { return m_HA_driver; }
+    
 protected:
     /// Internal method to init specific info. Called by init
-    virtual void initImpl() {}
-    
-    /// Main method to start haptic threads
-    virtual bool createHapticThreads() = 0;
+    virtual void initDevice() = 0;
     
     /// Main method to clear the device
-    void clearDevice();
+    virtual void clearDevice() = 0;
 
-    /// Main method to update the tool device from haptic information. 
-    /// Will call @sa updatePortalAnglesAndLength and @sa updatePositionImpl
-    virtual void updatePosition();
+    /// Main method from the SOFA simulation call at each simulation step begin.
+    virtual void simulation_updateData() = 0;
 
-    /// Method to propage
-    void updatePortalAnglesAndLength(sofa::type::fixed_array<float, 4> values);
-
-    /// Internal method to bo overriden by child class to propagate specific position. Called by @sa updatePosition
-    virtual void updatePositionImpl() = 0;
-
+    
     /// Internal method to bo overriden by child class to draw specific information. Called by @sa draw
-    virtual void drawImpl(const sofa::core::visual::VisualParams*) {};
+    virtual void drawImpl(const sofa::core::visual::VisualParams* vparams) { SOFA_UNUSED(vparams); }
 
     /// Internal method to bo overriden by child class to draw debug information. Called by @sa draw if d_drawDebug is true
-    virtual void drawDebug(const sofa::core::visual::VisualParams* vparams);    
+    virtual void drawDebug(const sofa::core::visual::VisualParams* vparams) { SOFA_UNUSED(vparams); }
+
+    /// Internal metod called by destructor to clear object.
+    void clearImpl();
 
 public:
     /// Name of the port for this device
@@ -81,64 +68,12 @@ public:
     /// Data to store Information received by HW device
     Data<std::string> d_hapticIdentity;
 
-
-    /// Data parameter to draw dof axis
-    Data<bool> d_drawDeviceAxis;
     /// Data parameter to draw debug information
-    Data<bool> d_drawDebug;
-    /// Data parameter to draw output logs
-    Data<bool> d_drawLogOutputs;
-
-    
-    /// Link to the portalManager component
-    SingleLink<HapticAvatar_BaseDeviceController, HapticAvatar_PortalManager, BaseLink::FLAG_STOREPATH | BaseLink::FLAG_STRONGLINK> l_portalMgr;
-
-public: 
-    /// Data public for haptic thread
-
-    /// Structure used to transfer data fromt he haptic thread to the simulation thread.
-    struct DeviceData
-    {
-        sofa::type::fixed_array<float, 4> anglesAndLength;
-        sofa::type::fixed_array<float, 4> motorValues;
-        sofa::type::fixed_array<float, 3> collisionForces;
-        int toolId;
-        float jawOpening;
-    };
-
-    /// Data belonging to the haptic thread only
-    DeviceData m_hapticData;
-    /// Data used in the copy thread to copy @sa m_hapticData into this data that can be used by simulation thread.
-    DeviceData m_simuData;
-
-    /// Boolean to warn scheduler when SOFA has started the simulation (changed by AnimateBeginEvent)
-    bool m_simulationStarted; 
-    
-    /// Bool to notify thread to stop work
-    std::atomic<bool> m_terminate;
-
+    Data<bool> d_drawDebug;    
 
 protected:
-    /// Pointer to the internal Driver for device API communication
-    HapticAvatar_DriverPort * m_HA_driver;
-    /// Pointer to the portal manager to get information from the current portal
-    HapticAvatar_PortalManager * m_portalMgr;
-    
-    /// values returned by tool: Rot angle, Pitch angle, z Length, Yaw Angle
-    DeviceData m_debugData;
-
-    /// Parameter to know if device is ready or not.
-    bool m_deviceReady;
-    /// Id of the port returned by portalManager
-    int m_portId;
-    
-    std::thread copy_thread;
-
-
-    sofa::type::Mat3x3f m_toolRot;
-    sofa::type::Mat3x3f m_toolRotInv;
-    sofa::type::Mat3x3f m_PortalRot;
-    sofa::type::Mat4x4f m_instrumentMtx;
+    /// Internal parameter to know if device is ready or not.
+    bool m_deviceReady = false;
 };
 
 } // namespace sofa::HapticAvatar
